requirements.txt
# Streamlit for the UI
streamlit>=1.24.0

# LangChain core library
langchain>=0.1.0

# Mistral AI integration (LangChain Mistral)
langchain-mistralai>=0.0.4

# Optional dependencies for memory & prompts (usually included with langchain)
pydantic>=1.10.0

# for .env
python-dotenv

# pathlib is part of Python standard library, no need to install
